+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/text/completion/ContentAssistPreferences_Default.java
@@ -12,9 +12,7 @@ package melnorme.lang.ide.ui.text.completion;
 
 import static melnorme.lang.ide.ui.LangUIPlugin.PLUGIN_ID;
 
-import melnorme.lang.ide.core.operations.ToolchainPreferences;
 import melnorme.lang.ide.core.utils.prefs.BooleanPreference;
-import melnorme.lang.ide.core.utils.prefs.IProjectPreference;
 import melnorme.lang.ide.core.utils.prefs.IntPreference;
 import melnorme.lang.ide.ui.ContentAssistPreferences;
 import melnorme.lang.ide.ui.LangUIPlugin;
@@ -36,12 +34,6 @@ public interface ContentAssistPreferences_Default {
 			new BooleanPreference(PLUGIN_ID, "AUTO_ACTIVATE__AlphaNumericTrigger", false);
 	IntPreference AUTO_ACTIVATE__Delay = 
 			new IntPreference(PLUGIN_ID, "AUTO_ACTIVATE__Delay", 200);
-	
-	
-	/* FIXME: remove this preference */
-	IProjectPreference<Boolean> ShowDialogIfContentAssistErrors = 
-			new BooleanPreference(PLUGIN_ID, "daemon_showDialogIfErrors", true,
-				ToolchainPreferences.USE_PROJECT_SETTINGS).getProjectPreference();
 	
 	
 	class Helper {