+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/editor/actions/GoFmtOperation.java
@@ -10,21 +10,28 @@
  *******************************************************************************/
 package com.googlecode.goclipse.ui.editor.actions;
 
+import static melnorme.utilbox.core.Assert.AssertNamespace.assertNotNull;
 import static melnorme.utilbox.core.CoreUtil.list;
 
 import java.nio.file.Path;
 
 import org.eclipse.ui.texteditor.ITextEditor;
 
+import com.googlecode.goclipse.core.GoProjectEnvironment;
 import com.googlecode.goclipse.core.GoToolPreferences;
 import com.googlecode.goclipse.tooling.env.GoEnvironment;
 
+import melnorme.lang.ide.ui.editor.actions.AbstractEditorToolOperation;
+import melnorme.lang.tooling.common.ops.IOperationMonitor;
+import melnorme.lang.tooling.toolchain.ops.ToolResponse;
 import melnorme.utilbox.collections.Indexable;
+import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.misc.Location;
 import melnorme.utilbox.misc.MiscUtil;
+import melnorme.utilbox.process.ExternalProcessHelper.ExternalProcessResult;
 
-public class GoFmtOperation extends AbstractEditorGoToolOperation {
+public class GoFmtOperation extends AbstractEditorToolOperation<String> {
 	
 	protected static final String RUN_GOFMT_OpName = "Run 'gofmt'";
 	
@@ -32,7 +39,16 @@ public class GoFmtOperation extends AbstractEditorGoToolOperation {
 		super(RUN_GOFMT_OpName, editor);
 	}
 	
+	protected ProcessBuilder pb;
+	
 	@Override
+	public void prepareOperation() throws CommonException {
+		super.prepareOperation();
+		
+		GoEnvironment goEnv = GoProjectEnvironment.getGoEnvironment(project);
+		pb = prepareProcessBuilder(goEnv);
+	}
+	
 	protected ProcessBuilder prepareProcessBuilder(GoEnvironment goEnv) throws CommonException {
 		Path gofmt = getGofmtLocation(goEnv);
 		Indexable<String> cmd = list(gofmt.toString());
@@ -48,6 +64,25 @@ public class GoFmtOperation extends AbstractEditorGoToolOperation {
 	
 	public static Location getGofmtLocationFromGoRoot(Location rootLocation) throws CommonException {
 		return rootLocation.resolve_fromValid("bin/gofmt" + MiscUtil.getExecutableSuffix());
+	}
+	
+	@Override
+	protected ToolResponse<String> doBackgroundValueComputation(IOperationMonitor om)
+			throws CommonException, OperationCancellation {
+		
+		ExternalProcessResult processResult = getToolService().runProcess(pb, getSource(), om);
+		if(processResult.exitValue != 0) {
+			return ToolResponse.newError(processResult.getStdErrBytes().toString());
+		}
+		
+		return new ToolResponse<>(processResult.getStdOutBytes().toString());
+	}
+	
+	@Override
+	protected void handleResultData(String resultData) throws CommonException {
+		assertNotNull(resultData);
+		
+		setEditorTextPreservingCarret(resultData);
 	}
 	
 }